See #257 and #258 for background.

I see a huge improvement in startup time, from >200ms to ~80ms, matching with zgen's startup time if zgen itself is sourced.

Comparison:
zgen(load cache file only): 40ms
zgen(load cache file and source zgen.zsh): 80ms
zplug(with this patch): 80ms
zplug(without this patch): 200ms