@pirj 2016-09-15
@jonathye good job! `init.zsh` dropped from 300ms to 140ms in my case.