@pirj 2016-09-15
Moreover, not reading those files *at all* only reduces startup by 10ms, so I would also drop this line:

```
# TODO: /^.*desc(ription)?: ?/
```