@NigoroJr 2016-09-22
@jonathye Thanks for your contribution! Love this ❤️ 

@pirj I'm not quite sure what you mean by that. If there's a further improvement you can make, could you send us another P-R?