@seaneking 2016-09-24
Any chance you could bump zplug's version number now this is merged? Coming from zgen the current perf is killing me.